requirements.txt
# Core Dependencies
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
scipy>=1.7.0

# Data Processing
pillow>=8.3.0
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Network Communication
socket
pickle

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0
tensorboard>=2.6.0

# Optional: For advanced visualization
plotly>=5.3.0

# Development Tools (Optional)
pytest>=6.2.0
black>=21.7b0
flake8>=3.9.0

